+++ src/pages/account/signup.js
@@ -6,23 +6,23 @@ export const SignUp = () => {
             <div className="container product-navbar">
                 <Navbar />
             </div>
-            <div class="account-page">
-                <div class="account-container">
-                    <div class="signUp-offer">
-                        <img src="../assets/vivir-models.png" alt=""></img>
-                        <div class="offer-info">
+            <div className="account-page">
+                <div className="account-container">
+                    <div className="signUp-offer">
+                        <img src="../assets/vivir-models.png" alt="signup-offer-img"></img>
+                        <div className="offer-info">
                             <p><strong>Flat Rs.400 + free shipping</strong></p>
                             <p>On your first order</p>
                         </div>
                     </div>
-                    <div class="form-container">
+                    <div className="form-container">
                         <AccLinks />
-                        <form class="register-form">
+                        <form className="register-form">
                             <input type="text" placeholder="Username" required />
                             <input type="email" placeholder="Email" required />
                             <input type="password" placeholder="Password" required />
                             <input type="password" placeholder="Confirm password" required />
-                            <button type="submit" class="account-btn">Signup</button>
+                            <button type="submit" className="account-btn">Signup</button>
                         </form>
                     </div>
                 </div>
